Change Hazelcast tests to use Testcontainers where possible
