+++ cache-hazelcast/src/test/groovy/io/micronaut/cache/hazelcast/HazelcastClientSyncCacheSpec.groovy
@@ -15,54 +15,33 @@
  */
 package io.micronaut.cache.hazelcast
 
-import com.hazelcast.config.Config
-import com.hazelcast.config.EvictionPolicy
-import com.hazelcast.config.MapConfig
-import com.hazelcast.config.MaxSizeConfig
-import com.hazelcast.core.Hazelcast
 import com.hazelcast.core.HazelcastInstance
-import io.micronaut.cache.SyncCache
 import io.micronaut.cache.tck.AbstractSyncCacheSpec
 import io.micronaut.context.ApplicationContext
-import spock.lang.Retry
+import org.testcontainers.containers.GenericContainer
+import org.testcontainers.spock.Testcontainers
 import spock.lang.Shared
 
 /**
  * @author Nirav Assar
  * @since 1.0
  */
-class HazelcastClientSyncCacheSpec extends AbstractSyncCacheSpec {
+@Testcontainers
+class HazelcastClientSyncCacheSpec extends AbstractSyncCacheSpec implements HazelcastClientSupport {
+
+    @Shared
+    GenericContainer hazelcast = new GenericContainer("hazelcast/hazelcast")
+            .withExposedPorts(5701)
 
     @Shared
     HazelcastInstance hazelcastServerInstance
 
-    def setupSpec() {
-        MapConfig mapConfig = new MapConfig()
-                .setMaxSizeConfig(new MaxSizeConfig()
-                        .setMaxSizePolicy(MaxSizeConfig.MaxSizePolicy.PER_PARTITION)
-                        .setSize(3))
-                .setEvictionPolicy(EvictionPolicy.LRU)
-                .setName("test")
-        Config config = new Config("sampleCache")
-        config.setProperty("hazelcast.partition.count", "1")
-        config.addMapConfig(mapConfig)
-        hazelcastServerInstance = Hazelcast.getOrCreateHazelcastInstance(config)
-    }
-
-    def cleanupSpec() {
-        hazelcastServerInstance.shutdown()
-    }
 
     @Override
     ApplicationContext createApplicationContext() {
         return ApplicationContext.run(
-                "hazelcast.client.instanceName": "sampleCache",
-                "hazelcast.client.network.addresses": ['127.0.0.1:5701']
+                "hazelcast.client.network.addresses": ["127.0.0.1:${hazelcast.firstMappedPort}"]
         )
-    }
-
-    @Override
-    void flushCache(SyncCache syncCache) {
     }
 
 }
