+++ cache-hazelcast/src/test/groovy/io/micronaut/cache/hazelcast/HazelcastMemberSyncCacheSpec.groovy
@@ -56,8 +56,4 @@ class HazelcastMemberSyncCacheSpec extends AbstractSyncCacheSpec {
         return ApplicationContext.run()
     }
 
-    @Override
-    void flushCache(SyncCache syncCache) {
-    }
-
 }
